feat: Move DukeException to exception package
