+++ src/main/java/command/AddDeadlineCommand.java
@@ -1,6 +1,6 @@
 package command;
 
-import duke.DukeException;
+import exception.DukeException;
 import task.Deadline;
 import task.TaskList;
 
@@ -28,7 +28,7 @@ public class AddDeadlineCommand implements Command {
     @Override
     public void execute(TaskList taskList) {
         Deadline deadline = new Deadline(taskDescription, by);
-        taskList.add(deadline);  // Add to the task list
+        taskList.add(deadline); // Add to the task list
         System.out.println("     Got it. I've added this task:");
         System.out.println("       " + deadline);
         System.out.println("     Now you have " + taskList.size() + " tasks in the list.");
