+++ src/main/java/command/AddEventCommand.java
@@ -1,6 +1,6 @@
 package command;
 
-import duke.DukeException;
+import exception.DukeException;
 import task.Event;
 import task.TaskList;
 
@@ -31,7 +31,7 @@ public class AddEventCommand implements Command {
     @Override
     public void execute(TaskList taskList) {
         Event event = new Event(taskDescription, from, to);
-        taskList.add(event);  // Add to the task list
+        taskList.add(event); // Add to the task list
         System.out.println("     Got it. I've added this task:");
         System.out.println("       " + event);
         System.out.println("     Now you have " + taskList.size() + " tasks in the list.");
